WEEK13/Daily_challenge2/requirements.txt
streamlit
langchain
langgraph
langchain-core
langchain-community
langchainhub
ipykernel
langchain_groq
beautifulsoup4
tiktoken
chromadb
langchain_google_genai
python-dotenv
ipywidgets
nbformat
nbconvert